+++ tests/testthat/test-datasets.R
@@ -4,13 +4,27 @@
 ## https://github.com/waldronlab/curatedTCGAData/blob/master/tests/testthat/test-resources.R
 
 test_that("metadata numbers match ExperimentHub", {
-    assays_file <- system.file("extdata", 
-                               "metadata.csv",
-                               package = "scpdata",
-                               mustWork = TRUE)
-    metadataFile <- read.csv(assays_file, 
-                             stringsAsFactors = FALSE)
+    metaf <- system.file("extdata", "metadata.csv", 
+                         package = "scpdata", mustWork = TRUE)
+    meta <- read.csv(metaf, stringsAsFactors = FALSE)
     EHub <- query(ExperimentHub(), "scpdata")
-    expect_equal(nrow(metadataFile), length(EHub))
-    expect_equal(metadataFile$Title, EHub$title)
+    expect_identical(nrow(meta), length(EHub))
+    expect_identical(meta$Title, EHub$title)
 })
+
+test_that("scpdata", {
+    metaf <- system.file("extdata", "metadata.csv", 
+                         package = "scpdata", mustWork = TRUE)
+    meta <- read.csv(metaf, stringsAsFactors = FALSE)
+    res <- scpdata()
+    expect_identical(meta$Title, res$title)
+})
+
+test_that("all datasets are available", {
+    res <- scpdata()
+    for (dataset in res$title) {
+        ds <- eval(call(dataset))
+        expect_true(is(ds, "QFeatures"))
+    }
+})
+